requirements.txt
anyio==3.6.2
APScheduler==3.10.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
loguru==0.6.0
msgpack==1.0.5
multidict==6.0.4
nonebot-adapter-onebot==2.2.3
nonebot-adapter-telegram==0.1.0b10
nonebot-plugin-apscheduler==0.2.0
nonebot2==2.0.0rc4
Pillow==9.5.0
pydantic==1.10.7
pygtrie==2.5.0
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.29.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
yarl==1.9.2
